Feat: Add Task Manager feature with process list UI
Introduces a new Task Manager page with process listing, global resource stats, and ability to end non-vital processes in order to win. Includes HTML, SCSS, and TypeScript files for both the main manager and individual task rows, as well as a minor update to the global color scheme in styles.scss.